requirements.txt
celery==5.3.4
requests==2.31.0
python-dotenv==1.0.0
